reformatting and moving things around
There was this really irritating table in the first
poppr paper that cited all of the references for
each function. This is nice if you aren't writing
a dissertation, but is a pain in the ass to format
if you want to have a nice latex or HTML
presentation. This was the solution